input unit testing in login-page part 1
